refactor: split env configs into client and server
