+++ lib/prisma.ts
@@ -1,5 +1,5 @@
 import { PrismaClient } from '@prisma/client'
-import { SERVER_ENV } from '@/lib/env'
+import { serverEnv } from '@/lib/env.server'
 
 // Cache Prisma Client instance on `globalThis` to avoid re-instantiation
 // during development hot reloads. Prisma connects lazily on first use.
@@ -9,6 +9,6 @@ const globalForPrisma = globalThis as unknown as {
 
 export const prisma = globalForPrisma.prisma ?? new PrismaClient()
 
-if (SERVER_ENV.NODE_ENV !== 'production') globalForPrisma.prisma = prisma
+if (serverEnv.NODE_ENV !== 'production') globalForPrisma.prisma = prisma
 
 export default prisma
